7.Deploying a Machine Learning Service with Flask and Django - Bank Loan Prediction/Project Service CNN/apiCNN/serializers.py
from rest_framework import serializers
from apiCNN import models


class ImagenSerializer(serializers.ModelSerializer):
    class Meta:
        fields = (
            'id',
            'name',
            'animal',
            'percentage'
        )
        model = models.Imagen

