7.Deploying a Machine Learning Service with Flask and Django - Bank Loan Prediction/Project Service CNN/requirements.txt
pip3 install tensorflow
pip3 install keras
pip3 install Django
pip3 install djangorestframework
pip3 install drf-yasg
pip3 install numpy
pip3 install pandas
pip3 install sklearn
pip3 install gunicorn
pip3 install pillow
pip3 install polls
pip3 install markdown
pip3 install django-filter
numpy~=1.18.1
pandas~=1.0.3
keras~=2.3.1
tensorflow~=2.1.0
django~=3.0.6
djangorestframework~=3.11.0